move BaseScoreboard into ScoreboardImpl
